--- app/UP/page.tsx
@@ -1,7 +1,0 @@
-export default function Page(){
-    return (
-        <main>
-            
-        </main>
-    );
-}
